Updated jest config and database functionality to fix testing errors

+++ my-portfolio/src/lib/database/c_stats.ts
@@ -1,4 +1,4 @@
-import { get_db, Collection } from '@src/lib/database/mongodb';
+import { get_db, Collection, get_db_name, Database } from '@src/lib/database/mongodb';
 import { ObjectId } from 'mongodb';
 
 //https://orangematter.solarwinds.com/2019/12/22/what-is-mongodbs-id-field-and-how-to-use-it/
@@ -58,7 +58,11 @@ export async function delete_stat(entry_id: string) {
     return result;
 }
 
-export async function clear_stats() {
+export async function clear_stats(force: boolean = false) {
+    if (get_db_name() == Database.portfolio && !force) {
+        throw new Error(`Cannot clear stats in ${Database.portfolio} without force being true`);
+    }
+
     const db = await get_db();
 
     const result = await db.collection<Stat>(Collection.stats).deleteMany({});
